+++ packages/substrate-client/src/chainhead/body.ts
@@ -3,7 +3,7 @@ import type { OperationBodyDoneRpc } from "./json-rpc-types"
 import { createOperationPromise } from "./operation-promise"
 
 export const createBodyFn = createOperationPromise(
-  () => chainHead.body,
+  chainHead.body,
   (hash: string) => [
     [hash],
     (e: OperationBodyDoneRpc, res: (x: Array<string>) => void) => {